requirements.txt
Flask==3.1.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
numpy==2.3.2
Pillow==11.3.0
python-dotenv==1.1.1
qrcode==8.2
SQLAlchemy==2.0.43
tensorflow==2.20.0
WTForms==3.2.1
